layout/style/test/test_bug499655.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=499655
-->
<head>
  <title>Test for Bug 499655</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=499655">Mozilla Bug 499655</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
<![CDATA[

test1 = document.createElementNS("http://www.w3.org/1999/xhtml","test");
test2 = document.createElementNS("http://www.w3.org/1999/xhtml","TEst");
test3 = document.createElementNS("test","test");
test4 = document.createElementNS("test","TEst");

content = document.getElementById("content");

content.appendChild(test1);
content.appendChild(test2);
content.appendChild(test3);
content.appendChild(test4);


list = document.querySelectorAll('test');
is(list.length, 2, "Number of elements found");
is(list[0], test1, "First element didn't match");
is(list[1] , test3, "Third element didn't match");

list = document.querySelectorAll('TEst');
is(list.length, 2, "Number of elements found");
is(list[0], test2, "Second element didn't match");
is(list[1], test4, "Fourth element didn't match");


]]>
</script>
</pre>
</body>
</html>
